Add leave room functionality
